# Conflicts:
#	drivers/ethernet/eth_stm32_hal.c
#	drivers/ethernet/phy/CMakeLists.txt
#	drivers/timer/cortex_m_systick.c
#	include/zephyr/kernel/thread.h
#	include/zephyr/net/net_core.h
#	include/zephyr/net/net_if.h
#	kernel/sched.c
#	subsys/net/ip/net_if.c
#	subsys/net/ip/net_private.h
#	subsys/net/ip/tcp.c
#	subsys/shell/shell.c
#	west.yml

+++ drivers/usb/udc/udc_stm32.c
@@ -414,7 +414,7 @@ static void handle_msg_setup(struct udc_stm32_data *priv)
 
 	if ((setup->bmRequestType == 0) && (setup->bRequest == USB_SREQ_SET_ADDRESS)) {
 		/* HAL requires we set the address before submitting status */
-		HAL_PCD_SetAddress(&priv->pcd, setup->wValue);
+		HAL_PCD_SetAddress(&priv->pcd, (uint8_t)setup->wValue);
 	}
 
 	if (udc_ctrl_stage_is_data_out(dev)) {
